+++ database/migrations/2022_05_07_084646_create_payments_table.php
@@ -15,9 +15,9 @@ class CreatePaymentsTable extends Migration
     {
         Schema::create('payments', function (Blueprint $table) {
             $table->id();
-            $table->float('amount')->comment('Amount of each fee');
-            $table->float('paid_amount')->comment('Amount paid from each fee');
-            $table->float('remaining')->comment('amount - paid amount');
+            $table->double('amount')->comment('Amount of each fee');
+            $table->double('paid_amount')->comment('Amount paid from each fee');
+            $table->double('remaining')->comment('amount - paid amount');
             $table->foreignId('studentsRegistration_id')->constrained('studentsRegistrations')->onUpdate('cascade')->comment('from students registration table is it ***id***');
             $table->foreignId('cours_fee_id')->constrained('cours_fees')->onUpdate('cascade')->comment('from cours fees ***id***');
             $table->foreignId('receipt_id')->constrained('receipts')->onUpdate('cascade')->comment('from from receipt ***id***');
